Start working on a possible prop pox patch

+++ SC4Fix/version.cpp
@@ -38,7 +38,7 @@ uint64_t GetAssemblyVersion(HMODULE hModule) {
 			&& size > 0) {
 			VS_FIXEDFILEINFO* verInfo = (VS_FIXEDFILEINFO*)lpBuffer;
 			if (verInfo->dwSignature = 0xfeef04bd) {
-				uint64_t qwValue = verInfo->dwFileVersionMS << 32;
+				uint64_t qwValue = (uint64_t)verInfo->dwFileVersionMS << 32;
 				qwValue |= verInfo->dwFileVersionLS;
 
 				return qwValue;
